+++ test/filter_test.dart
@@ -18,11 +18,20 @@ void main() {
 
     test('fillRect', () {
       Image f = Image.from(image);
-      int c = getColor(128, 255, 128, 255);
-      fillRect(f, 50, 50, 150, 150, c);
+      fillRect(f, 50, 50, 150, 150, getColor(128, 255, 128, 255));
+      fillRect(f, 250, -10, 100, 750, getColor(255, 128, 128, 128));
       File fp = File('out/fillRect.jpg');
       fp.createSync(recursive: true);
       fp.writeAsBytesSync(writeJpg(f));
+    });
+
+    test('floodFill', () {
+      Image s = readJpg(new File('test/res/oblique.jpg').readAsBytesSync());
+      int c = s.getPixel(50, 50);
+      fillFlood(s, 50, 50, c, threshold: 15.6);
+      File fp = File('out/fillFlood.jpg');
+      fp.createSync(recursive: true);
+      fp.writeAsBytesSync(writeJpg(s));
     });
 
     test('copyRectify', () {
@@ -126,9 +135,27 @@ void main() {
       Image f = Image.from(image);
       colorOffset(f, red: 5);
 
-      File fp = File('out/colorOffset.jpg');
-      fp.createSync(recursive: true);
-      fp.writeAsBytesSync(writeJpg(f));
+      File('out/colorOffset.jpg')
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(writeJpg(f));
+
+      f = Image(48, 48);
+      colorOffset(f, red: 255);
+      File('out/colorOffset_red.jpg')
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(writeJpg(f));
+
+      f = Image(48, 48);
+      colorOffset(f, green: 255);
+      File('out/colorOffset_green.jpg')
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(writeJpg(f));
+
+      f = Image(48, 48);
+      colorOffset(f, blue: 255);
+      File('out/colorOffset_blue.jpg')
+        ..createSync(recursive: true)
+        ..writeAsBytesSync(writeJpg(f));
     });
 
     test('contrast', () {